Driver Progress and Et Cetera
Started progress on the driver, mostly writing support code to set up simulation.
Added the Schedule class, which is a wrapper that holds a Truck object and the interval at which it arrives at the warehouse.
Added a null constructor to Truck, which randomly generates all relevant values. Names are drawn from a class pulled off the internet for this express purpose.