Generation methods & XML Comments
Created some new methods to help with things that need random generation in a few of the classes. I changed the constructors in these classes accordingly in order to use these methods for the ids and prices. I also filled in all the XML comments I could Find in the 4 main methods.